# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Description of ARIS backend API",
        "title": "ARIS backend API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/feed": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Getting feed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feed"
                ],
                "summary": "Get feed",
                "operationId": "get-feed",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "number of posts",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "cursor string responded by feed request",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/feed.FeedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.CommonErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/dto.CommonErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.CommonErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/posts/popular": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Getting popular posts",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feed"
                ],
                "summary": "Get popular posts",
                "operationId": "get-popular-users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/feed.popularPostsResponse"
                        }
                    }
                }
            }
        },
...
        "/public/feed": {
            "get": {
                "description": "Getting public feed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feed"
                ],
                "summary": "Get public feed",
                "operationId": "get-public-feed",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "number of posts",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "cursor string responded by feed request",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/feed.FeedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.CommonErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/dto.CommonErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.CommonErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CommonErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "feed.FeedResponse": {
            "type": "object",
            "properties": {
                "hasMore": {
                    "type": "boolean"
                },
                "nextCursor": {
                    "type": "string"
                },
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/feed.postFeedDTO"
                    }
                }
            }
        },
        "feed.authorFeedDTO": {
            "type": "object",
            "properties": {
                "avatarLink": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "feed.mediaFeedDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "mediaLink": {
                    "type": "string"
                },
                "mimeType": {
                    "type": "string"
                }
            }
        },
        "feed.popularPostDTO": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                }
            }
        },
        "feed.popularPostsResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/feed.popularPostDTO"
                    }
                }
            }
        },
        "feed.postFeedDTO": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/feed.authorFeedDTO"
                },
                "comments": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "likes": {
                    "type": "integer"
                },
                "medias": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/feed.mediaFeedDTO"
                    }
                },
                "reposts": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "SessionAuth": {
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    }
}